update oas2 Algorand4.4.1.2.2025072914

+++ dotnet-algorand-sdk/Algod/Model/Block.cs
@@ -165,7 +165,7 @@ namespace Algorand.Algod.Model
     [field:InspectorName(@"TxnCounter")]  
     public int TxnCounter { get; set; }
 #else
-        public int? TxnCounter { get; set; }
+        public ulong? TxnCounter { get; set; }
 #endif
